refactor(전체): 패키지 정비
