+++ src/main/java/bowling/controller/BowlingGameController.java
@@ -1,7 +1,7 @@
 package bowling.controller;
 
-import bowling.domain.PinCount;
-import bowling.domain.Player;
+import bowling.domain.state.PinCount;
+import bowling.domain.player.Player;
 import bowling.view.InputView;
 import bowling.view.OutputView;
