+++ src/main/java/bowling/domain/frame/Frame.java
@@ -1,6 +1,6 @@
 package bowling.domain.frame;
 
-import bowling.domain.PinCount;
+import bowling.domain.state.PinCount;
 import bowling.domain.state.State;
 import bowling.domain.state.StateHistory;
