+++ src/main/java/bowling/domain/state/End.java
@@ -1,7 +1,5 @@
 package bowling.domain.state;
 
-import bowling.domain.PinCount;
-
 public class End extends Finished {
 
     public static final String GAME_END_ERROR = "게임이 완료되었습니다. 플레이할 수 없습니다.";
