+++ src/main/java/bowling/domain/state/Gutter.java
@@ -1,7 +1,5 @@
 package bowling.domain.state;
 
-import bowling.domain.PinCount;
-
 import static bowling.Constants.MIN_FELLED_PIN_COUNT;
 
 public class Gutter extends NotFinished {
