+++ src/main/java/bowling/domain/state/PinCount.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.state;
 
 import static bowling.Constants.*;
