+++ src/main/java/bowling/domain/state/State.java
@@ -1,7 +1,5 @@
 package bowling.domain.state;
 
-import bowling.domain.PinCount;
-
 public interface State {
     State play(PinCount newFelledPin);
