+++ src/main/java/bowling/view/OutputView.java
@@ -1,7 +1,7 @@
 package bowling.view;
 
 import bowling.ViewUtils;
-import bowling.domain.Player;
+import bowling.domain.player.Player;
 import bowling.domain.frame.Frame;
 import bowling.domain.state.State;
