+++ src/test/java/bowling/domain/frame/FramesTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static bowling.domain.PinCountTest.*;
+import static bowling.domain.state.PinCountTest.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class FramesTest {
